+++ shared/util/tf/tf_generator.go
@@ -113,6 +113,10 @@ var requiredVariables []variable = []variable{
 
 const requiredTerraformVersion = ">= 1.3"
 
+var tagsAttrs = []string{"TFC_RUN_ID", "TFC_WORKSPACE_NAME", "TFC_WORKSPACE_SLUG", "TFC_CONFIGURATION_VERSION_GIT_BRANCH",
+	"TFC_CONFIGURATION_VERSION_GIT_COMMIT_SHA", "TFC_CONFIGURATION_VERSION_GIT_TAG", "TFC_PROJECT_NAME",
+	"project", "env", "service", "owner"}
+
 type TfGenerator struct {
 	happyConfig *config.HappyConfig
 }
@@ -394,21 +398,18 @@ func (tf TfGenerator) generateAwsProvider(rootBody *hclwrite.Body, alias, accoun
 	assumeRoleBlockBody.SetAttributeRaw("role_arn", tokens(fmt.Sprintf("\"arn:aws:iam::%s:role/%s\"", accountIdExpr, roleExpr)))
 
 	defaultTagsBlockBody := awsProviderBody.AppendNewBlock("default_tags", nil).Body()
-	tagsBlockBody := defaultTagsBlockBody.AppendNewBlock("tags", nil).Body()
 
-	tagsAttrs := [...]string{"TFC_RUN_ID", "TFC_WORKSPACE_NAME", "TFC_WORKSPACE_SLUG", "TFC_CONFIGURATION_VERSION_GIT_BRANCH",
-		"TFC_CONFIGURATION_VERSION_GIT_COMMIT_SHA", "TFC_CONFIGURATION_VERSION_GIT_TAG", "TFC_PROJECT_NAME",
-		"project", "env", "service", "owner"}
+	attrs := []hclwrite.ObjectAttrTokens{}
+
 	for _, tagAttr := range tagsAttrs {
-		s := ""
-		if strings.HasPrefix(tagAttr, "TFC_") {
-			s = fmt.Sprintf("coalesce(var.%s, \"unknown\")", tagAttr)
-		} else {
-			s = fmt.Sprintf("coalesce(var.tags.%s, \"unknown\")", tagAttr)
-		}
-		tagsBlockBody.SetAttributeRaw(tagAttr, tokens(s))
+		s := fmt.Sprintf("coalesce(var.%s, \"unknown\")", tagAttr)
+
+		attrs = append(attrs, hclwrite.ObjectAttrTokens{Name: tokens(tagAttr), Value: tokens(s)})
 	}
-	tagsBlockBody.SetAttributeRaw("managedBy", tokens("terraform"))
+	attrs = append(attrs, hclwrite.ObjectAttrTokens{Name: tokens("managedBy"), Value: tokens("\"terraform\"")})
+
+	tks := hclwrite.TokensForObject(attrs)
+	defaultTagsBlockBody.SetAttributeRaw("tags", tks)
 
 	awsProviderBody.SetAttributeRaw("allowed_account_ids", tokens(fmt.Sprintf("[\"%s\"]", accountIdExpr)))
 	return nil
@@ -501,6 +502,14 @@ func (tf *TfGenerator) GenerateVariables(srcDir string) error {
 		if !variable.Default.IsNull() {
 			variableBody.SetAttributeValue("default", variable.Default)
 		}
+	}
+	for _, tag := range tagsAttrs {
+		variableBody := rootBody.AppendNewBlock("variable", []string{tag}).Body()
+		tokens := hclwrite.TokensForTraversal(hcl.Traversal{
+			hcl.TraverseRoot{Name: "string"},
+		})
+		variableBody.SetAttributeRaw("type", tokens)
+		variableBody.SetAttributeValue("default", cty.StringVal("unknown"))
 	}
 
 	_, err = tfFile.Write(hclFile.Bytes())