+++ terraform/modules/happy-service-eks/main.tf
@@ -3,6 +3,9 @@ data "aws_region" "current" {}
 locals {
   tags_string  = join(",", [for key, val in local.routing_tags : "${key}=${val}"])
   service_type = (var.routing.service_type == "PRIVATE" || var.routing.service_mesh) ? "ClusterIP" : "NodePort"
+  match_labels = {
+    app = var.routing.service_name
+  }
   labels = merge({
     app                            = var.routing.service_name
     "app.kubernetes.io/name"       = var.stack_name
@@ -10,6 +13,24 @@ locals {
     "app.kubernetes.io/part-of"    = var.stack_name
     "app.kubernetes.io/managed-by" = "happy"
   }, var.additional_pod_labels)
+
+  external_dns_exclude_annotation = {
+    "external-dns.alpha.kubernetes.io/exclude" = "true"
+  }
+
+  base_ingress_variables = {
+    ingress_name            = var.routing.service_name
+    target_service_port     = var.routing.service_mesh ? 443 : var.routing.service_port
+    target_service_name     = var.routing.service_mesh ? "nginx-ingress-ingress-nginx-controller" : var.routing.service_name
+    target_service_scheme   = var.routing.service_mesh ? "HTTPS" : var.routing.service_scheme
+    cloud_env               = var.cloud_env
+    k8s_namespace           = var.routing.service_mesh ? "nginx-encrypted-ingress" : var.k8s_namespace
+    certificate_arn         = var.certificate_arn
+    tags_string             = local.tags_string
+    labels                  = local.labels
+    regional_wafv2_arn      = var.regional_wafv2_arn
+    ingress_security_groups = var.ingress_security_groups
+  }
 }
 
 resource "kubernetes_deployment_v1" "deployment" {
@@ -48,9 +69,7 @@ resource "kubernetes_deployment_v1" "deployment" {
     }
 
     selector {
-      match_labels = {
-        app = var.routing.service_name
-      }
+      match_labels = local.match_labels
     }
 
     template {
@@ -79,9 +98,7 @@ resource "kubernetes_deployment_v1" "deployment" {
 
       spec {
         service_account_name = var.aws_iam.service_account_name == null ? module.iam_service_account.service_account_name : var.aws_iam.service_account_name
-        node_selector = merge({
-          "kubernetes.io/arch" = var.gpu != null ? "amd64" : var.platform_architecture
-        }, var.gpu != null ? { "nvidia.com/gpu.present" = "true" } : {}, var.additional_node_selectors)
+
         dynamic "toleration" {
           for_each = var.gpu != null ? [1] : []
           content {
@@ -96,6 +113,51 @@ resource "kubernetes_deployment_v1" "deployment" {
             key      = "nvidia.com/gpu"
             operator = "Exists"
             effect   = "NoSchedule"
+          }
+        }
+
+        topology_spread_constraint {
+          max_skew = 3
+          #TODO: Once min_domains are supported, uncomment line below. https://github.com/hashicorp/terraform-provider-kubernetes/issues/2292
+          #min_domains        = 3
+          topology_key       = "topology.kubernetes.io/zone"
+          when_unsatisfiable = "DoNotSchedule"
+          label_selector {
+            match_labels = local.match_labels
+          }
+        }
+
+        affinity {
+          node_affinity {
+            required_during_scheduling_ignored_during_execution {
+              node_selector_term {
+                match_expressions {
+                  key      = "kubernetes.io/arch"
+                  operator = "In"
+                  values   = [var.platform_architecture]
+                }
+              }
+            }
+          }
+          pod_anti_affinity {
+            preferred_during_scheduling_ignored_during_execution {
+              weight = 100
+              pod_affinity_term {
+                topology_key = "kubernetes.io/hostname"
+                label_selector {
+                  match_labels = local.match_labels
+                }
+              }
+            }
+            preferred_during_scheduling_ignored_during_execution {
+              weight = 100
+              pod_affinity_term {
+                topology_key = "topology.kubernetes.io/zone"
+                label_selector {
+                  match_labels = local.match_labels
+                }
+              }
+            }
           }
         }
 
@@ -223,6 +285,7 @@ resource "kubernetes_deployment_v1" "deployment" {
 
             initial_delay_seconds = var.initial_delay_seconds
             period_seconds        = var.period_seconds
+            timeout_seconds       = var.liveness_timeout_seconds
           }
 
           readiness_probe {
@@ -234,6 +297,7 @@ resource "kubernetes_deployment_v1" "deployment" {
 
             initial_delay_seconds = var.initial_delay_seconds
             period_seconds        = var.period_seconds
+            timeout_seconds       = var.readiness_timeout_seconds
           }
         }
 
@@ -269,6 +333,7 @@ resource "kubernetes_deployment_v1" "deployment" {
 
               initial_delay_seconds = container.value.initial_delay_seconds
               period_seconds        = container.value.period_seconds
+              timeout_seconds       = container.value.liveness_timeout_seconds
             }
 
             readiness_probe {
@@ -280,6 +345,7 @@ resource "kubernetes_deployment_v1" "deployment" {
 
               initial_delay_seconds = container.value.initial_delay_seconds
               period_seconds        = container.value.period_seconds
+              timeout_seconds       = container.value.readiness_timeout_seconds
             }
 
             dynamic "volume_mount" {
@@ -412,29 +478,51 @@ resource "kubernetes_service_v1" "service" {
 }
 
 module "ingress" {
-  count = (var.routing.service_type == "EXTERNAL" || var.routing.service_type == "INTERNAL" || var.routing.service_type == "VPC") ? 1 : 0
+  count  = (var.routing.service_type == "EXTERNAL" || var.routing.service_type == "INTERNAL" || var.routing.service_type == "VPC") ? 1 : 0
+  source = "../happy-ingress-eks"
 
-  source                  = "../happy-ingress-eks"
-  ingress_name            = var.routing.service_name
-  target_service_port     = var.routing.service_mesh ? 443 : var.routing.service_port
-  target_service_name     = var.routing.service_mesh ? "nginx-ingress-ingress-nginx-controller" : var.routing.service_name
-  target_service_scheme   = var.routing.service_mesh ? "HTTPS" : var.routing.service_scheme
-  cloud_env               = var.cloud_env
-  k8s_namespace           = var.routing.service_mesh ? "nginx-encrypted-ingress" : var.k8s_namespace
-  certificate_arn         = var.certificate_arn
-  tags_string             = local.tags_string
-  routing                 = var.routing
-  labels                  = local.labels
-  regional_wafv2_arn      = var.regional_wafv2_arn
-  ingress_security_groups = var.ingress_security_groups
+  ingress_name            = local.base_ingress_variables.ingress_name
+  target_service_port     = local.base_ingress_variables.target_service_port
+  target_service_name     = local.base_ingress_variables.target_service_name
+  target_service_scheme   = local.base_ingress_variables.target_service_scheme
+  cloud_env               = local.base_ingress_variables.cloud_env
+  k8s_namespace           = local.base_ingress_variables.k8s_namespace
+  certificate_arn         = local.base_ingress_variables.certificate_arn
+  tags_string             = local.base_ingress_variables.tags_string
+  labels                  = local.base_ingress_variables.labels
+  regional_wafv2_arn      = local.base_ingress_variables.regional_wafv2_arn
+  ingress_security_groups = local.base_ingress_variables.ingress_security_groups
+
+  routing = var.routing
+}
+
+module "ingress_exclude_external_dns" {
+  for_each = (var.routing.service_type == "EXTERNAL" || var.routing.service_type == "INTERNAL" || var.routing.service_type == "VPC") ? var.routing.additional_hostnames : []
+  source   = "../happy-ingress-eks"
+
+  ingress_name            = replace("${local.base_ingress_variables.ingress_name}-${each.value}", ".", "-")
+  target_service_port     = local.base_ingress_variables.target_service_port
+  target_service_name     = local.base_ingress_variables.target_service_name
+  target_service_scheme   = local.base_ingress_variables.target_service_scheme
+  cloud_env               = local.base_ingress_variables.cloud_env
+  k8s_namespace           = local.base_ingress_variables.k8s_namespace
+  certificate_arn         = local.base_ingress_variables.certificate_arn
+  tags_string             = local.base_ingress_variables.tags_string
+  labels                  = local.base_ingress_variables.labels
+  regional_wafv2_arn      = local.base_ingress_variables.regional_wafv2_arn
+  ingress_security_groups = local.base_ingress_variables.ingress_security_groups
+
+  routing                = merge(var.routing, { host_match : "" })
+  additional_annotations = local.external_dns_exclude_annotation
 }
 
 module "nginx-ingress" {
-  count               = ((var.routing.service_type == "EXTERNAL" || var.routing.service_type == "INTERNAL" || var.routing.service_type == "VPC") && var.routing.service_mesh) ? 1 : 0
-  source              = "../happy-nginx-ingress-eks"
-  ingress_name        = "${var.routing.service_name}-nginx"
+  for_each = ((var.routing.service_type == "EXTERNAL" || var.routing.service_type == "INTERNAL" || var.routing.service_type == "VPC") && var.routing.service_mesh) ? setunion([var.routing.host_match], var.routing.additional_hostnames) : []
+  source   = "../happy-nginx-ingress-eks"
+
+  ingress_name        = replace("${var.routing.service_name}-${each.value}-nginx", ".", "-")
   k8s_namespace       = var.k8s_namespace
-  host_match          = var.routing.host_match
+  host_match          = each.value
   host_path           = replace(var.routing.path, "/\\*$/", "") //NGINX does not support paths that end with *
   target_service_name = var.routing.service_name
   target_service_port = var.routing.service_port
@@ -472,6 +560,22 @@ resource "kubernetes_horizontal_pod_autoscaler_v1" "hpa" {
       api_version = "apps/v1"
       kind        = "Deployment"
       name        = kubernetes_deployment_v1.deployment[0].metadata[0].name
+    }
+  }
+}
+
+resource "kubernetes_pod_disruption_budget_v1" "pdb" {
+  count = var.routing.service_type == "IMAGE_TEMPLATE" || var.max_unavailable_count >= var.desired_count ? 0 : 1
+  metadata {
+    name      = var.routing.service_name
+    namespace = var.k8s_namespace
+    labels    = local.labels
+  }
+
+  spec {
+    max_unavailable = var.max_unavailable_count
+    selector {
+      match_labels = local.match_labels
     }
   }
 }