+++ terraform/modules/happy-cloudfront/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.1.0](https://github.com/chanzuckerberg/happy/compare/happy-cloudfront-v1.0.1...happy-cloudfront-v1.1.0) (2023-11-02)
+
+
+### Features
+
+* separate configuration of allow methods and allowed cache methods ([#2667](https://github.com/chanzuckerberg/happy/issues/2667)) ([af4518a](https://github.com/chanzuckerberg/happy/commit/af4518ac1cc90096294b2bb1c629e62db2f6b700))
+
+## [1.0.1](https://github.com/chanzuckerberg/happy/compare/happy-cloudfront-v1.0.0...happy-cloudfront-v1.0.1) (2023-10-31)
+
+
+### Bug Fixes
+
+* redirect HTTP traffic to HTTPS from viewer to CloudFront ([#2665](https://github.com/chanzuckerberg/happy/issues/2665)) ([fd95af9](https://github.com/chanzuckerberg/happy/commit/fd95af94e710d05becd2a769f6afe5c3be2ee532))
+
 ## 1.0.0 (2023-10-03)